refactor: 멀티 모듈 구조에 맞춰 bootstrap 레이어 분리
