+++ nmnb-bootstrap/src/main/resources/application.yml
@@ -1,0 +1,7 @@
+spring:
+  application:
+    name: nmnb
+  servlet:
+    multipart:
+      max-file-size: 100MB
+      max-request-size: 100MB
